kusto-ingestion/requirements.txt
azure-kusto-data>=4.4
pandas>=2.2
azure-identity>=1.16
